stm32/timer: merge pwm module into timer.
The traits there are applicable to timer use cases other than PWM.
It doesn't make sense to keep them separated.
