+++ embassy-stm32/src/timer/mod.rs
@@ -1,3 +1,6 @@
+pub mod complementary_pwm;
+pub mod simple_pwm;
+
 use stm32_metapac::timer::vals;
 
 use crate::interrupt;
@@ -43,7 +46,86 @@ pub(crate) mod sealed {
     pub trait AdvancedControlInstance: GeneralPurpose16bitInstance {
         fn regs_advanced() -> crate::pac::timer::TimAdv;
     }
+
+    pub trait CaptureCompare16bitInstance: GeneralPurpose16bitInstance {
+        /// Global output enable. Does not do anything on non-advanced timers.
+        fn enable_outputs(&mut self, enable: bool);
+
+        fn set_output_compare_mode(&mut self, channel: Channel, mode: OutputCompareMode);
+
+        fn enable_channel(&mut self, channel: Channel, enable: bool);
+
+        fn set_compare_value(&mut self, channel: Channel, value: u16);
+
+        fn get_max_compare_value(&self) -> u16;
+    }
+
+    pub trait ComplementaryCaptureCompare16bitInstance: CaptureCompare16bitInstance {
+        fn set_dead_time_clock_division(&mut self, value: vals::Ckd);
+
+        fn set_dead_time_value(&mut self, value: u8);
+
+        fn enable_complementary_channel(&mut self, channel: Channel, enable: bool);
+    }
+
+    pub trait CaptureCompare32bitInstance: GeneralPurpose32bitInstance {
+        fn set_output_compare_mode(&mut self, channel: Channel, mode: OutputCompareMode);
+
+        fn enable_channel(&mut self, channel: Channel, enable: bool);
+
+        fn set_compare_value(&mut self, channel: Channel, value: u32);
+
+        fn get_max_compare_value(&self) -> u32;
+    }
 }
+
+#[derive(Clone, Copy)]
+pub enum Channel {
+    Ch1,
+    Ch2,
+    Ch3,
+    Ch4,
+}
+
+impl Channel {
+    pub fn raw(&self) -> usize {
+        match self {
+            Channel::Ch1 => 0,
+            Channel::Ch2 => 1,
+            Channel::Ch3 => 2,
+            Channel::Ch4 => 3,
+        }
+    }
+}
+
+#[derive(Clone, Copy)]
+pub enum OutputCompareMode {
+    Frozen,
+    ActiveOnMatch,
+    InactiveOnMatch,
+    Toggle,
+    ForceInactive,
+    ForceActive,
+    PwmMode1,
+    PwmMode2,
+}
+
+impl From<OutputCompareMode> for stm32_metapac::timer::vals::Ocm {
+    fn from(mode: OutputCompareMode) -> Self {
+        match mode {
+            OutputCompareMode::Frozen => stm32_metapac::timer::vals::Ocm::FROZEN,
+            OutputCompareMode::ActiveOnMatch => stm32_metapac::timer::vals::Ocm::ACTIVEONMATCH,
+            OutputCompareMode::InactiveOnMatch => stm32_metapac::timer::vals::Ocm::INACTIVEONMATCH,
+            OutputCompareMode::Toggle => stm32_metapac::timer::vals::Ocm::TOGGLE,
+            OutputCompareMode::ForceInactive => stm32_metapac::timer::vals::Ocm::FORCEINACTIVE,
+            OutputCompareMode::ForceActive => stm32_metapac::timer::vals::Ocm::FORCEACTIVE,
+            OutputCompareMode::PwmMode1 => stm32_metapac::timer::vals::Ocm::PWMMODE1,
+            OutputCompareMode::PwmMode2 => stm32_metapac::timer::vals::Ocm::PWMMODE2,
+        }
+    }
+}
+
+pub trait Basic16bitInstance: sealed::Basic16bitInstance + 'static {}
 
 pub trait GeneralPurpose16bitInstance: sealed::GeneralPurpose16bitInstance + 'static {}
 
@@ -51,7 +133,36 @@ pub trait GeneralPurpose32bitInstance: sealed::GeneralPurpose32bitInstance + 'st
 
 pub trait AdvancedControlInstance: sealed::AdvancedControlInstance + 'static {}
 
-pub trait Basic16bitInstance: sealed::Basic16bitInstance + 'static {}
+pub trait CaptureCompare16bitInstance:
+    sealed::CaptureCompare16bitInstance + GeneralPurpose16bitInstance + 'static
+{
+}
+
+pub trait ComplementaryCaptureCompare16bitInstance:
+    sealed::ComplementaryCaptureCompare16bitInstance + AdvancedControlInstance + 'static
+{
+}
+
+pub trait CaptureCompare32bitInstance:
+    sealed::CaptureCompare32bitInstance + CaptureCompare16bitInstance + GeneralPurpose32bitInstance + 'static
+{
+}
+
+pin_trait!(Channel1Pin, CaptureCompare16bitInstance);
+pin_trait!(Channel1ComplementaryPin, CaptureCompare16bitInstance);
+pin_trait!(Channel2Pin, CaptureCompare16bitInstance);
+pin_trait!(Channel2ComplementaryPin, CaptureCompare16bitInstance);
+pin_trait!(Channel3Pin, CaptureCompare16bitInstance);
+pin_trait!(Channel3ComplementaryPin, CaptureCompare16bitInstance);
+pin_trait!(Channel4Pin, CaptureCompare16bitInstance);
+pin_trait!(Channel4ComplementaryPin, CaptureCompare16bitInstance);
+pin_trait!(ExternalTriggerPin, CaptureCompare16bitInstance);
+pin_trait!(BreakInputPin, CaptureCompare16bitInstance);
+pin_trait!(BreakInputComparator1Pin, CaptureCompare16bitInstance);
+pin_trait!(BreakInputComparator2Pin, CaptureCompare16bitInstance);
+pin_trait!(BreakInput2Pin, CaptureCompare16bitInstance);
+pin_trait!(BreakInput2Comparator1Pin, CaptureCompare16bitInstance);
+pin_trait!(BreakInput2Comparator2Pin, CaptureCompare16bitInstance);
 
 #[allow(unused)]
 macro_rules! impl_basic_16bit_timer {
@@ -140,63 +251,116 @@ macro_rules! impl_32bit_timer {
     };
 }
 
+#[allow(unused)]
+macro_rules! impl_compare_capable_16bit {
+    ($inst:ident) => {
+        impl sealed::CaptureCompare16bitInstance for crate::peripherals::$inst {
+            fn enable_outputs(&mut self, _enable: bool) {}
+
+            fn set_output_compare_mode(&mut self, channel: Channel, mode: OutputCompareMode) {
+                use sealed::GeneralPurpose16bitInstance;
+                let r = Self::regs_gp16();
+                let raw_channel: usize = channel.raw();
+                r.ccmr_output(raw_channel / 2)
+                    .modify(|w| w.set_ocm(raw_channel % 2, mode.into()));
+            }
+
+            fn enable_channel(&mut self, channel: Channel, enable: bool) {
+                use sealed::GeneralPurpose16bitInstance;
+                Self::regs_gp16()
+                    .ccer()
+                    .modify(|w| w.set_cce(channel.raw(), enable));
+            }
+
+            fn set_compare_value(&mut self, channel: Channel, value: u16) {
+                use sealed::GeneralPurpose16bitInstance;
+                Self::regs_gp16().ccr(channel.raw()).modify(|w| w.set_ccr(value));
+            }
+
+            fn get_max_compare_value(&self) -> u16 {
+                use sealed::GeneralPurpose16bitInstance;
+                Self::regs_gp16().arr().read().arr()
+            }
+        }
+    };
+}
+
 foreach_interrupt! {
     ($inst:ident, timer, TIM_BASIC, UP, $irq:ident) => {
         impl_basic_16bit_timer!($inst, $irq);
-
-        impl Basic16bitInstance for crate::peripherals::$inst {
-        }
+        impl Basic16bitInstance for crate::peripherals::$inst {}
     };
     ($inst:ident, timer, TIM_GP16, UP, $irq:ident) => {
         impl_basic_16bit_timer!($inst, $irq);
-
-        impl Basic16bitInstance for crate::peripherals::$inst {
-        }
+        impl_compare_capable_16bit!($inst);
+        impl Basic16bitInstance for crate::peripherals::$inst {}
+        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {}
+        impl CaptureCompare16bitInstance for crate::peripherals::$inst {}
 
         impl sealed::GeneralPurpose16bitInstance for crate::peripherals::$inst {
             fn regs_gp16() -> crate::pac::timer::TimGp16 {
                 crate::pac::$inst
             }
         }
-
-        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {
-        }
     };
 
     ($inst:ident, timer, TIM_GP32, UP, $irq:ident) => {
         impl_basic_16bit_timer!($inst, $irq);
+        impl_32bit_timer!($inst);
+        impl_compare_capable_16bit!($inst);
+        impl Basic16bitInstance for crate::peripherals::$inst {}
+        impl CaptureCompare16bitInstance for crate::peripherals::$inst {}
+        impl CaptureCompare32bitInstance for crate::peripherals::$inst {}
+        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {}
+        impl GeneralPurpose32bitInstance for crate::peripherals::$inst {}
 
-        impl Basic16bitInstance for crate::peripherals::$inst {
+        impl sealed::CaptureCompare32bitInstance for crate::peripherals::$inst {
+            fn set_output_compare_mode(
+                &mut self,
+                channel: Channel,
+                mode: OutputCompareMode,
+            ) {
+                use crate::timer::sealed::GeneralPurpose32bitInstance;
+                let raw_channel = channel.raw();
+                Self::regs_gp32().ccmr_output(raw_channel / 2).modify(|w| w.set_ocm(raw_channel % 2, mode.into()));
+            }
+
+            fn enable_channel(&mut self, channel: Channel, enable: bool) {
+                use crate::timer::sealed::GeneralPurpose32bitInstance;
+                Self::regs_gp32().ccer().modify(|w| w.set_cce(channel.raw(), enable));
+            }
+
+            fn set_compare_value(&mut self, channel: Channel, value: u32) {
+                use crate::timer::sealed::GeneralPurpose32bitInstance;
+                Self::regs_gp32().ccr(channel.raw()).modify(|w| w.set_ccr(value));
+            }
+
+            fn get_max_compare_value(&self) -> u32 {
+                use crate::timer::sealed::GeneralPurpose32bitInstance;
+                Self::regs_gp32().arr().read().arr() as u32
+            }
         }
 
         impl sealed::GeneralPurpose16bitInstance for crate::peripherals::$inst {
             fn regs_gp16() -> crate::pac::timer::TimGp16 {
                 unsafe { crate::pac::timer::TimGp16::from_ptr(crate::pac::$inst.as_ptr()) }
             }
-        }
-
-        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {
-        }
-
-        impl_32bit_timer!($inst);
-
-        impl GeneralPurpose32bitInstance for crate::peripherals::$inst {
         }
     };
 
     ($inst:ident, timer, TIM_ADV, UP, $irq:ident) => {
         impl_basic_16bit_timer!($inst, $irq);
 
-        impl Basic16bitInstance for crate::peripherals::$inst {
-        }
+        impl Basic16bitInstance for crate::peripherals::$inst {}
+        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {}
+        impl CaptureCompare16bitInstance for crate::peripherals::$inst {}
+        impl ComplementaryCaptureCompare16bitInstance for crate::peripherals::$inst {}
+        impl AdvancedControlInstance for crate::peripherals::$inst {}
 
         impl sealed::GeneralPurpose16bitInstance for crate::peripherals::$inst {
             fn regs_gp16() -> crate::pac::timer::TimGp16 {
                 unsafe { crate::pac::timer::TimGp16::from_ptr(crate::pac::$inst.as_ptr()) }
             }
-        }
-
-        impl GeneralPurpose16bitInstance for crate::peripherals::$inst {
         }
 
         impl sealed::AdvancedControlInstance for crate::peripherals::$inst {
@@ -205,7 +369,64 @@ foreach_interrupt! {
             }
         }
 
-        impl AdvancedControlInstance for crate::peripherals::$inst {
+        impl sealed::CaptureCompare16bitInstance for crate::peripherals::$inst {
+            fn enable_outputs(&mut self, enable: bool) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                let r = Self::regs_advanced();
+                r.bdtr().modify(|w| w.set_moe(enable));
+            }
+
+            fn set_output_compare_mode(
+                &mut self,
+                channel: Channel,
+                mode: OutputCompareMode,
+            ) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                let r = Self::regs_advanced();
+                let raw_channel: usize = channel.raw();
+                r.ccmr_output(raw_channel / 2)
+                    .modify(|w| w.set_ocm(raw_channel % 2, mode.into()));
+            }
+
+            fn enable_channel(&mut self, channel: Channel, enable: bool) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced()
+                    .ccer()
+                    .modify(|w| w.set_cce(channel.raw(), enable));
+            }
+
+            fn set_compare_value(&mut self, channel: Channel, value: u16) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced()
+                    .ccr(channel.raw())
+                    .modify(|w| w.set_ccr(value));
+            }
+
+            fn get_max_compare_value(&self) -> u16 {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced().arr().read().arr()
+            }
         }
+
+        impl sealed::ComplementaryCaptureCompare16bitInstance for crate::peripherals::$inst {
+            fn set_dead_time_clock_division(&mut self, value: vals::Ckd) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced().cr1().modify(|w| w.set_ckd(value));
+            }
+
+            fn set_dead_time_value(&mut self, value: u8) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced().bdtr().modify(|w| w.set_dtg(value));
+            }
+
+            fn enable_complementary_channel(&mut self, channel: Channel, enable: bool) {
+                use crate::timer::sealed::AdvancedControlInstance;
+                Self::regs_advanced()
+                    .ccer()
+                    .modify(|w| w.set_ccne(channel.raw(), enable));
+            }
+        }
+
+
     };
 }
