+++ embassy-stm32/src/timer/simple_pwm.rs
