+++ examples/stm32h7/src/bin/low_level_timer_api.rs
@@ -6,8 +6,8 @@ use defmt::*;
 use embassy_executor::Spawner;
 use embassy_stm32::gpio::low_level::AFType;
 use embassy_stm32::gpio::Speed;
-use embassy_stm32::pwm::*;
 use embassy_stm32::time::{khz, mhz, Hertz};
+use embassy_stm32::timer::*;
 use embassy_stm32::{into_ref, Config, Peripheral, PeripheralRef};
 use embassy_time::{Duration, Timer};
 use {defmt_rtt as _, panic_probe as _};
